+++ tests/bootstrap.php
@@ -1,4 +1,6 @@
 <?php
 
-declare(strict_types=1);
+// phpcs:disable PSR12.Files.FileHeader.SpacingAfterTagBlock
+// phpcs:disable PSR1.Files.SideEffects.FoundWithSymbols
 
+declare(strict_types=1);